Change to use {{ }} for values
As it matches 'jinja' and we assume they do that for a reason and it
seems like a better fit when we start to introduce {% and %} which also
matches jinja.

There might be a better, nicer approach to all of this but for the
moment I'd rather struggle with the implementation than with the syntax
and we can refactor later.