Support 'import' and 'with' statements
These are somewhat hastily included statements to allow us to state the
type of certain parameters as Gleam doesn't infer partial record types.
ie. if we use 'user.is_admin' in an if-statement, it cannot infer that
user is some kind of object with at least a boolean 'is_admin' field.
Instead it asks for the type of the object.

We introduce the 'import' line to allow us to import definitions and
the 'with' line to specify the types of particular params.

We also add the '{>' syntax to indicate a statement that contains
everything on the rest of the line. Without this, if we had {% %}
around, say, the import statement then it wouldn't be clear how to
handle the new line at the end of the statement. Could it be part of the
text of the template or magically swallowed? Is annoys me when template
language force you to put arbitrary amounts of stuff onto a single line
in order to avoid new lines bleeding into the generated text.
Particularly annoying for generating plain text emails as it is nice to
control every space and new line in them.

With this syntax, the whole line is ignored and the template essentially
continues on the next line.