Simplify scanning
Instead of using knowledge of the keywords to guide the scanning we just
scan as many tokens as possible in each block and then let the parsing
error as needed. The scanning was doing too much before. Unfortunately
it cannot become completely stupid (I think) becase we need to
distinguish blocks from plain text so that we can scan the plain text as
single chunks whilst scanning the inside of blocks as keywords &
identifiers for the most part.

I'm not a expert so there might be a better way to approach this.