Provide coloured output for errors
We enabled coloured error messages when outputting to the terminal but
use the writer abstraction to use plain text for errors in test.

I've used ColorChoice::Auto but I can't get it to turn off by
redirecting through '| cat' or something but I might be missing
something. We can copy the approach from gleam at some point if needed.