Require all parameters to be explicitly declared
Using 'with' statements.

Rather than attempting to infer them from the usage in the template
itself. This has been suggested by @lpil and @michallepicki in order to
make it easier to import modules and use them within the templates and
avoid the engine attempting to infer the module name as a parameter or
generally failing in other ways.

There is an error if you declare the same thing twice using 'with'
statements.

This does mean that users need to explicitly declare all the parameters
name and declare types for String and StringBuilder values. We might be
able to remove the {[]} syntax for StringBuilder injection as we'll
already have that information from the 'with' statements.