Add 'DO NOT EDIT' comment to generated output
So that new users to a project know that that content is generated
automatically and they shouldn't edit it.

We embed the program name for the moment even though we'd like to change
it. We use CARGO_PKG_NAME so that it will automatically change when we
rename in Cargo.toml.