Support arbitrary tokens when scanning
Instead of requiring keywoards or valid identifiers after, say, an 'if'
we allow or arbitrary space separated tokens until we find the relevant
token that ends the current section, eg. '%}'.

This means that we can begin to support arbitrary gleam syntax in some
sections which gives the template writer more flexibility and they can
do some simple work in the template itself rather than having to
preprocess everything.

I've included a basic test that allows for 'items != []' in an if
condition.

I've also used an approach where we clone the iterator, move the clone
forward and both the original and the clone to check for pairs of tokens
like '%}' which allows us to be sure that we're encountering the end
token instead of some arbitrary bit of gleam code that starts with '%'
or something.