Only add gleam/list import if there is for-loop
If we don't have a for-loop in the template then it isn't guaranteed
that we'll use the gleam/list import which means that we make get
compiler warnings from the gleam build.

Here we track the use of for-loops and pass a boolean update the
rendering stack and use it to determine if we need the import.

As reported by Louis Pilfold.