Support in-template function declarations
I find it frustrating to work with text templates as it is often hard to
define helper functions without defining them in another module and
importing them in some fashion.

This is acceptable but can be annoying when wanting little components
for things like HTML rendering. Especially when using approaches like
tailwindcss.

I don't know what other templates system do this. I can't see support in
Django templates which is the one I'm most familiar with. It might be
avoided for a reason and I just don't have the foresight to undestand
why.

This adds support, tests and basic documentation.