Update to include 'type' in relevant imports
And bump dependencies which are quite old now.

I've removed 'import gleam/list' from one of the templates as we import
it anyway. It might be an issue if users want to import and use others
parts of gleam/list in their templates. It can probably be worked
around for the moment but something to keep in mind for the future.