Re-write requires_update logic to remove unwraps
And to use a direct comparison instead of the duration_since method
which I assume is ok.

Otherwise, we just and_then the results and exit early if we're not
getting access to the 'modified' times that we need to make the
decision. We conservatively require an update if we fail to access the
modified times.